requirements.txt
# ArtEmis Image Caption Generation - Dependencies

# Deep Learning (PyTorch)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0

# NLP
nltk>=3.8.0
gensim>=4.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0

# Utilities
tqdm>=4.65.0
